Serialise IntPtr lists/arrays as longs due to compat issues.
The client was crashing with messages like "The process was terminated due to an internal error in the .NET Runtime at IP 00007FFD19AAE1D8 (00007FFD19920000) with exit code 80131506.", which I've never seen before.

+++ src/PSADT/PSADT.ClientServer.Server/DataSerialization.cs
@@ -378,7 +378,7 @@ namespace PSADT.ClientServer
                 // DataContractSerializer needs concrete types, not interfaces
                 typeof(ReadOnlyCollection<string>),
                 typeof(ReadOnlyCollection<int>),
-                typeof(ReadOnlyCollection<nint>),
+                typeof(ReadOnlyCollection<long>),
                 typeof(ReadOnlyCollection<Hashtable>),
                 typeof(ReadOnlyCollection<ProcessDefinition>),
             ]
